moved model_metadata.py to constants folder
